requirements.txt
autopep8==2.3.2
flake8==7.3.0
mccabe==0.7.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
